+++ agent/src/main/java/com/walmartlabs/concord/agent/AgentConstants.java
@@ -4,14 +4,14 @@ package com.walmartlabs.concord.agent;
  * *****
  * Concord
  * -----
- * Copyright (C) 2017 - 2018 Walmart Inc.
+ * Copyright (C) 2017 - 2019 Walmart Inc.
  * -----
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,20 +20,12 @@ package com.walmartlabs.concord.agent;
  * =====
  */
 
+public final class AgentConstants {
 
-import com.google.inject.Guice;
-import com.google.inject.Injector;
-import org.eclipse.sisu.space.SpaceModule;
-import org.eclipse.sisu.space.URLClassSpace;
-import org.eclipse.sisu.wire.WireModule;
+    public static final long ERROR_DELAY = 5000;
+    public static final int API_CALL_MAX_RETRIES = 3;
+    public static final long API_CALL_RETRY_DELAY = 3000;
 
-public class Main {
-
-    public static void main(String[] args) throws Exception {
-        ClassLoader cl = Main.class.getClassLoader();
-        Injector injector = Guice.createInjector(new WireModule(new SpaceModule(new URLClassSpace(cl))));
-
-        Agent a = injector.getInstance(Agent.class);
-        a.run();
+    private AgentConstants() {
     }
 }