+++ agent/src/main/java/com/walmartlabs/concord/agent/DefaultStateFetcher.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.agent;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -26,29 +26,31 @@ import com.walmartlabs.concord.common.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.inject.Inject;
+import javax.inject.Named;
+import javax.inject.Singleton;
 import java.io.File;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
 
-public class RemoteStateFetcher implements Worker.StateFetcher {
+@Named
+@Singleton
+public class DefaultStateFetcher implements StateFetcher {
 
-    private static final Logger log = LoggerFactory.getLogger(RemoteStateFetcher.class);
+    private static final Logger log = LoggerFactory.getLogger(DefaultStateFetcher.class);
 
     private final ProcessApi processApi;
-    private final int maxRetries;
-    private final long retryDelay;
 
-    public RemoteStateFetcher(ProcessApi processApi, int maxRetries, long retryDelay) {
+    @Inject
+    public DefaultStateFetcher(ProcessApi processApi) {
         this.processApi = processApi;
-        this.maxRetries = maxRetries;
-        this.retryDelay = retryDelay;
     }
 
     @Override
     public void downloadState(JobRequest job) throws Exception {
         File payload = null;
         try {
-            payload = ClientUtils.withRetry(maxRetries, retryDelay, () -> processApi.downloadState(job.getInstanceId()));
+            payload = ClientUtils.withRetry(AgentConstants.API_CALL_MAX_RETRIES, AgentConstants.API_CALL_RETRY_DELAY, () -> processApi.downloadState(job.getInstanceId()));
             IOUtils.unzip(payload.toPath(), job.getPayloadDir(), StandardCopyOption.REPLACE_EXISTING);
         } finally {
             if (payload != null) {