+++ agent/src/main/java/com/walmartlabs/concord/agent/cfg/ConfigProvider.java
@@ -1,4 +1,4 @@
-package com.walmartlabs.concord.agent.executors;
+package com.walmartlabs.concord.agent.cfg;
 
 /*-
  * *****
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.agent.executors;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * 
  *      http://www.apache.org/licenses/LICENSE-2.0
- *
+ * 
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -20,12 +20,22 @@ package com.walmartlabs.concord.agent.executors;
  * =====
  */
 
-import com.walmartlabs.concord.agent.JobInstance;
-import com.walmartlabs.concord.agent.JobRequest;
+import com.typesafe.config.Config;
+import com.walmartlabs.ollie.config.ConfigurationProcessor;
+import com.walmartlabs.ollie.config.Environment;
+import com.walmartlabs.ollie.config.EnvironmentSelector;
 
-public interface JobExecutor {
+import javax.inject.Named;
+import javax.inject.Provider;
+import javax.inject.Singleton;
 
-    JobRequest.Type acceptsType();
+@Named
+@Singleton
+public class ConfigProvider implements Provider<Config> {
 
-    JobInstance exec(JobRequest jobRequest) throws Exception;
+    @Override
+    public Config get() {
+        Environment env = new EnvironmentSelector().select();
+        return new ConfigurationProcessor("concord-agent", env).process();
+    }
 }